added multiple views

+++ Blog.Logic/Services/CommentService.cs
@@ -3,59 +3,63 @@ using Blog.Data.Entities;
 using Blog.Data.Repositories;
 using Blog.Data.UoW;
 using Blog.Logic.Models;
+using Blog.Logic.Exceptions;
 
-namespace Blog.Logic.Services
+namespace Blog.Logic.Services;
+
+public class CommentService : ICommentService
 {
-    public class CommentService : ICommentService
+    private readonly IRepository<CommentEntity> _repo;
+    private readonly IMapper _mapper;
+
+    public CommentService(
+        IUnitOfWork unitOfWork,
+        IMapper mapper)
     {
-        private readonly IRepository<CommentEntity> _repo;
-        private readonly IMapper _mapper;
+        _repo = unitOfWork.GetRepository<CommentEntity>(false);
+        _mapper = mapper;
+    }
 
-        public CommentService(
-            IUnitOfWork unitOfWork,
-            IMapper mapper)
-        {
-            _repo = unitOfWork.GetRepository<CommentEntity>(false);
-            _mapper = mapper;
-        }
+    public async Task CreateComment(CommentModel comment)
+    {
+        var entity = _mapper.Map<CommentEntity>(comment);
 
-        public async Task CreateComment(CommentModel comment)
-        {
-            var entity = _mapper.Map<CommentEntity>(comment);
+        await _repo.Create(entity);
+    }
 
-            await _repo.Create(entity);
-        }
+    public async Task<CommentModel> GetComment(int id)
+    {
+        var entity = await _repo.Get(id);
+        var comment = _mapper.Map<CommentModel>(entity);
 
-        public async Task<CommentModel> GetComment(int id)
-        {
-            var entity = await _repo.Get(id);
-            var comment = _mapper.Map<CommentModel>(entity);
+        return comment;
+    }
 
-            return comment;
-        }
+    public List<CommentModel> GetComments()
+    {
+        var entities = _repo.GetAll();
+        var comments = _mapper.Map<List<CommentModel>>(entities);
 
-        public List<CommentModel> GetComments()
-        {
-            var entities = _repo.GetAll();
-            var comments = _mapper.Map<List<CommentModel>>(entities);
+        return comments;
+    }
 
-            return comments;
-        }
+    public async Task UpdateComment(CommentModel comment)
+    {
+        var entity = await _repo.Get(comment.Id);
 
-        public async Task UpdateComment(CommentModel comment)
-        {
-            var entity = await _repo.Get(comment.Id);
+        if (entity == null) throw new CommentNotFoundException();
 
-            entity.Content = comment.Content;
+        entity.Content = comment.Content;
 
-            await _repo.Update(entity);
-        }
+        await _repo.Update(entity);
+    }
 
-        public async Task DeleteComment(CommentModel comment)
-        {
-            var entity = await _repo.Get(comment.Id);
+    public async Task DeleteComment(CommentModel comment)
+    {
+        var entity = await _repo.Get(comment.Id);
 
-            await _repo.Delete(entity);
-        }
+        if (entity == null) throw new CommentNotFoundException();
+
+        await _repo.Delete(entity);
     }
 }
